We now support two different winding rules
Both non-zero and even-odd fill rules are now supported. They are
exposed by Batik anyways and we store them in the file format now.

Increased the file version to 3. Old files will be automatically
interpreted as non-zero fill rule everywhere. The old library may ignore
paths that are stored in the new format with the even-odd rule.